Split up chart.js before adding new chart types
